move internal ed25519 API into the aux namespace
