+++ src/ed25519/keypair.cpp
@@ -1,12 +1,12 @@
 // ignore warnings in this file
 #include "libtorrent/aux_/disable_warnings_push.hpp"
 
-#include "libtorrent/ed25519.hpp"
+#include "libtorrent/aux_/ed25519.hpp"
 #include "libtorrent/hasher512.hpp"
 #include "ge.h"
 
-namespace libtorrent
-{
+namespace libtorrent {
+namespace aux {
 
 void ed25519_create_keypair(unsigned char *public_key, unsigned char *private_key, const unsigned char *seed) {
     ge_p3 A;
@@ -21,4 +21,4 @@ void ed25519_create_keypair(unsigned char *public_key, unsigned char *private_ke
     ge_p3_tobytes(public_key, &A);
 }
 
-}
+} }
